+++ common/src/main/java/net/bettercombat/logic/PlayerAttackHelper.java
@@ -5,7 +5,6 @@ import net.bettercombat.api.AttackHand;
 import net.bettercombat.api.ComboState;
 import net.bettercombat.api.WeaponAttributes;
 import net.bettercombat.utils.AttributeModifierHelper;
-import net.minecraft.component.DataComponentTypes;
 import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
@@ -229,23 +228,32 @@ public class PlayerAttackHelper {
         return new Pose(mainPose, offPose);
     }
 
-    public static double getRange(PlayerEntity player, ItemStack stack) {
-        if (EntityAttributeHelper.itemHasRangeAttribute(stack)) {
-            return player.getEntityInteractionRange();
-        }
-        var attributes = WeaponRegistry.getAttributes(stack);
-        return getRange(player, attributes, true);
-    }
+
 
     public static double getStaticRange(PlayerEntity player, ItemStack stack) {
         var attributes = WeaponRegistry.getAttributes(stack);
-        return getRange(player, attributes, false);
+        return combineAttackRange(attributes, player.getAttributeBaseValue(EntityAttributes.PLAYER_ENTITY_INTERACTION_RANGE));
     }
 
-    public static double getRange(PlayerEntity player, WeaponAttributes attributes, boolean realValue) {
-        var range = realValue
-                ? player.getEntityInteractionRange()
-                : EntityAttributes.PLAYER_ENTITY_INTERACTION_RANGE.value().getDefaultValue();
+    public static double getRangeForItem(PlayerEntity player, ItemStack stack) {
+        var interactionRangeValue = player.getAttributeValue(EntityAttributes.PLAYER_ENTITY_INTERACTION_RANGE);
+        return getRangeWithItem(stack, interactionRangeValue);
+    }
+
+    public static double getRangeWithWeapon(PlayerEntity player, double interactionRangeValue) {
+        return getRangeWithItem(player.getMainHandStack(), interactionRangeValue);
+    }
+
+    private static double getRangeWithItem(ItemStack stack, double interactionRangeValue) {
+        if (EntityAttributeHelper.itemHasRangeAttribute(stack)) {
+            return interactionRangeValue;
+        }
+        var attributes = WeaponRegistry.getAttributes(stack);
+        return combineAttackRange(attributes, interactionRangeValue);
+    }
+
+    public static double combineAttackRange(WeaponAttributes attributes, double interactionRangeValue) {
+        var range = interactionRangeValue;
         if (attributes != null) {
             // Absolute range (legacy)
             if (attributes.attackRange() != 0) {