+++ common/src/main/java/net/bettercombat/mixin/client/AbstractClientPlayerEntityMixin.java
@@ -23,7 +23,7 @@ import net.bettercombat.client.animation.modifier.TransmissionSpeedModifier;
 import net.bettercombat.client.compat.FirstPersonAnimationCompatibility;
 import net.bettercombat.logic.AnimatedHand;
 import net.bettercombat.logic.PlayerAttackHelper;
-import net.bettercombat.mixin.LivingEntityAccessor;
+import net.bettercombat.mixin.player.LivingEntityAccessor;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.network.AbstractClientPlayerEntity;
 import net.minecraft.client.world.ClientWorld;