+++ common/src/main/java/net/bettercombat/mixin/player/LivingEntityMixin.java
@@ -1,4 +1,4 @@
-package net.bettercombat.mixin;
+package net.bettercombat.mixin.player;
 
 import net.bettercombat.logic.PlayerAttackHelper;
 import net.bettercombat.logic.PlayerAttackProperties;