+++ common/src/main/java/net/bettercombat/mixin/player/PlayerEntityAccessor.java
@@ -1,4 +1,4 @@
-package net.bettercombat.mixin;
+package net.bettercombat.mixin.player;
 
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;