+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.1.3
+
+- Improved compatibility with Entity Interaction Range attribute (`PlayerEntity.getEntityInteractionRange` returns a value with weapon attributes range bonus applied)
+
 # 2.1.2
 
 - Improved compatibility with Entity Interaction Range attribute